refactor(auth): simplify enabled/disabled JWT service configuration flow
